libsolid-rs: test login automatically, introduce client trait
Introduce initial support for automatically testing the login flow in
the client. This requires restructurung the login flow and introduce
custom requests on behalf of the user.

Moving some of the generic code from the `auth` into the root library
code, preparing for separation of logic for implementing the various
client features, as well as preparing for the deserializing and
serializing of client state.